@@ -1,28 +1,28 @@
 {
-  "_from": "@ngx-translate/core@12.1.2",
-  "_id": "@ngx-translate/core@12.1.2",
+  "_from": "@ngx-translate/core@13.0.0",
+  "_id": "@ngx-translate/core@13.0.0",
   "_inBundle": false,
-  "_integrity": "sha512-ZudJsqIxTKlLmPoqK8gJY3UpMGujR0Xm7HfXL6AR79yGRS23QqpjAhMfx4v5qUCcHMmQ9/78bW8QJLfp31c7vQ==",
+  "_integrity": "sha512-+tzEp8wlqEnw0Gc7jtVRAJ6RteUjXw6JJR4O65KlnxOmJrCGPI0xjV/lKRnQeU0w4i96PQs/jtpL921Wrb7PWg==",
   "_location": "/@ngx-translate/core",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "@ngx-translate/core@12.1.2",
+    "raw": "@ngx-translate/core@13.0.0",
     "name": "@ngx-translate/core",
     "escapedName": "@ngx-translate%2fcore",
     "scope": "@ngx-translate",
-    "rawSpec": "12.1.2",
+    "rawSpec": "13.0.0",
     "saveSpec": null,
-    "fetchSpec": "12.1.2"
+    "fetchSpec": "13.0.0"
   },
   "_requiredBy": [
     "#USER",
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/@ngx-translate/core/-/core-12.1.2.tgz",
-  "_shasum": "0c6f24249953a79cfc2d581b2cd1b5d6d338d9db",
-  "_spec": "@ngx-translate/core@12.1.2",
+  "_resolved": "https://registry.npmjs.org/@ngx-translate/core/-/core-13.0.0.tgz",
+  "_shasum": "60547cb8a0845a2a0abfde6b0bf5ec6516a63fd6",
+  "_spec": "@ngx-translate/core@13.0.0",
   "_where": "D:\\Onemoney\\Test Lib\\ngSLDemo\\projects\\my-lib",
   "author": {
     "name": "Olivier Combe"
@@ -31,13 +31,14 @@
     "url": "https://github.com/ngx-translate/core/issues"
   },
   "bundleDependencies": false,
+  "dependencies": {
+    "tslib": "^2.0.0"
+  },
   "deprecated": false,
   "description": "The internationalization (i18n) library for Angular",
   "es2015": "fesm2015/ngx-translate-core.js",
   "esm2015": "esm2015/ngx-translate-core.js",
-  "esm5": "esm5/ngx-translate-core.js",
   "fesm2015": "fesm2015/ngx-translate-core.js",
-  "fesm5": "fesm5/ngx-translate-core.js",
   "homepage": "https://github.com/ngx-translate/core",
   "keywords": [
     "angular",
@@ -49,12 +50,11 @@
   "license": "MIT",
   "main": "bundles/ngx-translate-core.umd.js",
   "metadata": "ngx-translate-core.metadata.json",
-  "module": "fesm5/ngx-translate-core.js",
+  "module": "fesm2015/ngx-translate-core.js",
   "name": "@ngx-translate/core",
   "peerDependencies": {
-    "@angular/core": ">=8.0.0",
-    "rxjs": ">=6.3.0",
-    "tslib": "^1.10.0"
+    "@angular/core": ">=10.0.0",
+    "rxjs": ">=6.5.3"
   },
   "repository": {
     "type": "git",
@@ -65,5 +65,5 @@
   },
   "sideEffects": false,
   "typings": "ngx-translate-core.d.ts",
-  "version": "12.1.2"
+  "version": "13.0.0"
 }
