@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { CustomDashboardComponent } from './custom-dashboard/custom-dashboard.component';
 
 import { HttpClient, HttpClientModule } from '@angular/common/http';
-import { OnemoneyWebsdkService } from 'onemoney-headless-websdk';
+import { MyLibService as OnemoneyWebsdkService } from '../lib/my-lib.service';
 
 import { RouterModule, Routes } from '@angular/router';
 import { AuthenticationComponent } from '../app/pages/authentication/authentication.component';
@@ -11,7 +11,6 @@ import { MessagesComponent } from '../app/components/popups/messages/messages.co
 import { ToastComponent } from '../app/components/toast/toast.component';
 import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
 //import { EmbedVideo } from 'ngx-embed-video';
-import { HttpLoaderFactory } from '../lib/main.module';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
  import { LoginComponent } from '../app/pages/login/login.component';
 import { FormComponent } from '../app/util-components/form/form.component';
@@ -31,14 +30,7 @@ import { FormComponent } from '../app/util-components/form/form.component';
     FormsModule,
     ReactiveFormsModule,
     //EmbedVideo.forRoot(),
-    TranslateModule.forRoot({
-      loader: {
-        provide: TranslateLoader,
-        useFactory: HttpLoaderFactory,
-        deps: [HttpClient]
-      },
-      defaultLanguage: 'en'
-    })
+    
     
   ],
   declarations: [
