@@ -4,7 +4,7 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 
 import { HttpClientModule, HttpClient } from '@angular/common/http';
-import { TranslateModule } from '@ngx-translate/core';
+import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 
 @NgModule({
@@ -24,12 +24,9 @@ import { TranslateHttpLoader } from '@ngx-translate/http-loader';
     BrowserModule,
     BrowserAnimationsModule,
     HttpClientModule,
-   
+    
   ],
   providers: []
 })
 export class MainModule { }
 
-export function HttpLoaderFactory(http: HttpClient) {
-  return new TranslateHttpLoader(http, '../assets/i18n/', '.json');
-}
